+++ src/parser.ts
@@ -4,6 +4,35 @@
 import * as vscode from 'vscode';
 import * as jsonc from 'jsonc-parser';
 import { DEFAULT_RANGE } from './diagnostics';
+import { EFFECT_ENTITY_REGEX, ENTITY_REGEX } from './regex';
+
+/*
+ * see https://code.visualstudio.com/api/language-extensions/semantic-highlight-guide
+ * see https://github.com/microsoft/node-jsonc-parser
+ */
+
+const tokenTypes = ['namespace', 'type', 'property', 'macro', 'function'];
+const tokenModifiers = ['declaration', 'deprecated'];
+export const semanticTokensLegend = new vscode.SemanticTokensLegend(
+  tokenTypes,
+  tokenModifiers
+);
+
+const determineDefinitionRange = (
+  type: string,
+  line: number,
+  start: number,
+  length: number
+) => {
+  const pos = Math.max(type.lastIndexOf('::'));
+  const typePos = pos === -1 ? 0 : pos + 2;
+  const defRange = new vscode.Range(
+    new vscode.Position(line, start + 1 + typePos),
+    new vscode.Position(line, start + length)
+  );
+
+  return defRange;
+};
 
 // Cedar policies
 
@@ -16,6 +45,8 @@ export type PolicyRange = {
 type PolicyCacheItem = {
   version: number;
   policies: PolicyRange[];
+  tokens: vscode.SemanticTokens;
+  definitions: vscode.Range[];
 };
 
 const policyCache: Record<string, PolicyCacheItem> = {};
@@ -24,22 +55,24 @@ const ID_ATTR = /@id\("(?<id>(.+))"\)/;
 
 export const parseCedarDocPolicies = (
   cedarDoc: vscode.TextDocument,
-  visitPolicy:
+  visitPolicy?:
     | undefined
     | ((policyRange: PolicyRange, policyText: string) => void)
-): PolicyRange[] => {
+): PolicyCacheItem => {
   // policy text is not cached, so check for no visitPolicy callback
   if (visitPolicy === undefined) {
     const cachedItem = policyCache[cedarDoc.uri.toString()];
     if (cachedItem && cachedItem.version === cedarDoc.version) {
       // console.log("parseCedarDocPolicies (cached)");
-      return cachedItem.policies;
+      return cachedItem;
     }
   }
 
   const policies: PolicyRange[] = [];
+  const definitions: vscode.Range[] = [];
   let count = 0;
   let id: string | null = null;
+  const tokensBuilder = new vscode.SemanticTokensBuilder(semanticTokensLegend);
 
   let tmpPolicy = '';
   let effectRange: vscode.Range = DEFAULT_RANGE;
@@ -73,12 +106,26 @@ export const parseCedarDocPolicies = (
     }
 
     const commentPos = textLine.indexOf('//');
-    if (
-      textLine
-        .substring(0, commentPos > -1 ? commentPos : textLine.length)
-        .trim()
-        .endsWith(';')
-    ) {
+    const nonCommentLine = textLine.substring(
+      0,
+      commentPos > -1 ? commentPos : textLine.length
+    );
+    let found = nonCommentLine.match(EFFECT_ENTITY_REGEX);
+    if (found && found?.groups) {
+      const type = found?.groups.type;
+      const startCharacter = nonCommentLine.indexOf(type) || 0;
+      const typeRange = new vscode.Range(
+        new vscode.Position(i, startCharacter),
+        new vscode.Position(i, startCharacter + type.length)
+      );
+      tokensBuilder.push(typeRange, 'type', []);
+
+      definitions.push(
+        determineDefinitionRange(type, i, startCharacter - 1, type.length + 1)
+      );
+    }
+
+    if (nonCommentLine.trim().endsWith(';')) {
       const policyRange = {
         id: id || `policy${count}`,
         range: new vscode.Range(
@@ -100,12 +147,23 @@ export const parseCedarDocPolicies = (
     }
   }
 
-  policyCache[cedarDoc.uri.toString()] = {
+  const cachedItem = {
     version: cedarDoc.version,
     policies: policies,
+    tokens: tokensBuilder.build(),
+    definitions: definitions,
   };
+  policyCache[cedarDoc.uri.toString()] = cachedItem;
 
-  return policies;
+  return cachedItem;
+};
+
+export const cedarTokensProvider: vscode.DocumentSemanticTokensProvider = {
+  provideDocumentSemanticTokens(
+    cedarDoc: vscode.TextDocument
+  ): vscode.ProviderResult<vscode.SemanticTokens> {
+    return parseCedarDocPolicies(cedarDoc).tokens;
+  },
 };
 
 // Cedar entities
@@ -122,29 +180,32 @@ export type EntityRange = {
   parentsRange: vscode.Range | null;
 };
 
-type EntityCacheItem = {
+export type EntityCacheItem = {
   version: number;
   entities: EntityRange[];
+  tokens: vscode.SemanticTokens;
+  definitions: vscode.Range[];
 };
 
 const entityCache: Record<string, EntityCacheItem> = {};
 
 export const parseCedarDocEntities = (
   entitiesDoc: vscode.TextDocument,
-  visitEntity:
+  visitEntity?:
     | undefined
     | ((entityRange: EntityRange, entityText: string) => void)
-): EntityRange[] => {
+): EntityCacheItem => {
   // entity text is not cached, so check for no visitEntity callback
   if (visitEntity === undefined) {
     const cachedItem = entityCache[entitiesDoc.uri.toString()];
     if (cachedItem && cachedItem.version === entitiesDoc.version) {
       // console.log("parseCedarDocEntities (cached)");
-      return cachedItem.entities;
+      return cachedItem;
     }
   }
 
   const entities: EntityRange[] = [];
+  const definitions: vscode.Range[] = [];
 
   let uidType: string = '';
   let uid: string = '';
@@ -161,6 +222,7 @@ export const parseCedarDocEntities = (
   let depth1Property = '';
   let attrsNameRanges: Record<string, vscode.Range> = {};
   let parentsCount = 0;
+  const tokensBuilder = new vscode.SemanticTokensBuilder(semanticTokensLegend);
 
   jsonc.visit(entitiesDoc.getText(), {
     onObjectBegin(offset, length, startLine, startCharacter, pathSupplier) {
@@ -234,30 +296,34 @@ export const parseCedarDocEntities = (
       startCharacter,
       pathSupplier
     ) {
-      if (pathSupplier().length === 1) {
+      const range = new vscode.Range(
+        new vscode.Position(startLine, startCharacter),
+        new vscode.Position(startLine, startCharacter + length)
+      );
+      const innerRange = new vscode.Range(
+        new vscode.Position(startLine, startCharacter + 1),
+        new vscode.Position(startLine, startCharacter + length - 1)
+      );
+      const jsonPathLen = pathSupplier().length;
+      if (jsonPathLen === 1) {
         depth1Property = property;
         if (property === 'uid') {
-          uidKeyRange = new vscode.Range(
-            new vscode.Position(startLine, startCharacter + 1),
-            new vscode.Position(startLine, startCharacter + length - 1)
-          );
+          uidKeyRange = innerRange;
         } else if (property === 'attrs') {
-          attrsKeyRange = new vscode.Range(
-            new vscode.Position(startLine, startCharacter + 1),
-            new vscode.Position(startLine, startCharacter + length - 1)
-          );
+          attrsKeyRange = innerRange;
         } else if (property === 'parents') {
-          parentsKeyRange = new vscode.Range(
-            new vscode.Position(startLine, startCharacter + 1),
-            new vscode.Position(startLine, startCharacter + length - 1)
-          );
+          parentsKeyRange = innerRange;
         }
-      }
-      if (pathSupplier().length === 2 && pathSupplier()[1] === 'attrs') {
-        attrsNameRanges[property] = new vscode.Range(
-          new vscode.Position(startLine, startCharacter + 1),
-          new vscode.Position(startLine, startCharacter + length - 1)
-        );
+      } else if (jsonPathLen === 2 && pathSupplier()[1] === 'attrs') {
+        // anything directly under "attrs" is an property
+        tokensBuilder.push(range, 'property', []);
+        attrsNameRanges[property] = innerRange;
+      } else if (property === '__expr') {
+        // treat "__expr" as a deprecated macro
+        tokensBuilder.push(range, 'macro', ['deprecated']);
+      } else if (property === '__entity' || property === '__extn') {
+        // treat "__entity" and "__extn as a macro
+        tokensBuilder.push(range, 'macro', []);
       }
     },
     onLiteralValue(
@@ -268,6 +334,10 @@ export const parseCedarDocEntities = (
       startCharacter,
       pathSupplier
     ) {
+      const range = new vscode.Range(
+        new vscode.Position(startLine, startCharacter),
+        new vscode.Position(startLine, startCharacter + length)
+      );
       const jsonPathLen = pathSupplier().length;
       if (pathSupplier()[1] === 'uid') {
         if (pathSupplier()[jsonPathLen - 1] === 'type') {
@@ -284,57 +354,117 @@ export const parseCedarDocEntities = (
         }
         if (pathSupplier()[jsonPathLen - 1] === 'uid') {
           uid = value;
+          let found = uid.match(ENTITY_REGEX);
+          if (found?.groups) {
+            const type = found?.groups.type;
+            const typeRange = new vscode.Range(
+              new vscode.Position(startLine, startCharacter + 1),
+              new vscode.Position(startLine, startCharacter + type.length + 1)
+            );
+            tokensBuilder.push(typeRange, 'type', []);
+
+            definitions.push(
+              determineDefinitionRange(
+                type,
+                startLine,
+                startCharacter,
+                type.length + 1
+              )
+            );
+          }
         }
+      }
+
+      if (
+        jsonPathLen > 2 &&
+        pathSupplier()[jsonPathLen - 1] === 'type' &&
+        (pathSupplier()[1] === 'uid' ||
+          pathSupplier()[1] === 'parents' ||
+          pathSupplier()[jsonPathLen - 2] === '__entity')
+      ) {
+        // most things directly under "type" are a type
+        tokensBuilder.push(range, 'type', []);
+
+        definitions.push(
+          determineDefinitionRange(value, startLine, startCharacter, length - 1)
+        );
+      } else if (
+        jsonPathLen > 2 &&
+        pathSupplier()[jsonPathLen - 1] === 'fn' &&
+        pathSupplier()[jsonPathLen - 2] === '__extn'
+      ) {
+        // things under "__extn" then "fn" are a function
+        tokensBuilder.push(range, 'function', []);
       }
     },
   });
 
-  entityCache[entitiesDoc.uri.toString()] = {
+  const cachedItem = {
     version: entitiesDoc.version,
     entities: entities,
+    tokens: tokensBuilder.build(),
+    definitions: definitions,
   };
+  entityCache[entitiesDoc.uri.toString()] = cachedItem;
 
-  return entities;
+  return cachedItem;
+};
+
+// analyze the Cedar entities JSON documents and return semantic tokens
+export const entitiesTokensProvider: vscode.DocumentSemanticTokensProvider = {
+  provideDocumentSemanticTokens(
+    cedarEntitiesDoc: vscode.TextDocument
+  ): vscode.ProviderResult<vscode.SemanticTokens> {
+    return parseCedarDocEntities(cedarEntitiesDoc).tokens;
+  },
 };
 
 // Cedar schema
 
 export type SchemaRange = {
   etype: string;
+  deftype: string;
   range: vscode.Range;
   etypeRange: vscode.Range;
+  symbol: vscode.SymbolKind;
 };
 
 type SchemaCacheItem = {
   version: number;
   entities: SchemaRange[];
+  tokens: vscode.SemanticTokens;
+  definitions: vscode.Range[];
 };
 
 const schemaCache: Record<string, SchemaCacheItem> = {};
 
 export const parseCedarDocSchema = (
   schemaDoc: vscode.TextDocument,
-  visitSchema:
+  visitSchema?:
     | undefined
     | ((schemaRange: SchemaRange, schemaText: string) => void)
-): SchemaRange[] => {
+): SchemaCacheItem => {
   // schema text is not cached, so check for no visitSchema callback
   if (visitSchema === undefined) {
     const cachedItem = schemaCache[schemaDoc.uri.toString()];
     if (cachedItem && cachedItem.version === schemaDoc.version) {
       // console.log("parseCedarDocSchema (cached)");
-      return cachedItem.entities;
+      return cachedItem;
     }
   }
 
   const entities: SchemaRange[] = [];
+  const definitions: vscode.Range[] = [];
 
   let namespace: string = '';
   let etype: string = '';
+  let deftype: string = '';
   let etypeStart: vscode.Position | null = null;
   let etypeEnd: vscode.Position | null = null;
   let etypeRange: vscode.Range | null = null;
   let depth = 0;
+  let symbol = vscode.SymbolKind.Class;
+  const tokensBuilder = new vscode.SemanticTokensBuilder(semanticTokensLegend);
 
   jsonc.visit(schemaDoc.getText(), {
     onObjectBegin(offset, length, startLine, startCharacter, pathSupplier) {
@@ -347,8 +477,10 @@ export const parseCedarDocSchema = (
         if (etypeStart && etypeEnd && etypeRange) {
           const schemaRange: SchemaRange = {
             etype: etype,
+            deftype: deftype,
             range: new vscode.Range(etypeStart, etypeEnd),
             etypeRange: etypeRange,
+            symbol: symbol,
           };
           entities.push(schemaRange);
         }
@@ -362,33 +494,123 @@ export const parseCedarDocSchema = (
       startCharacter,
       pathSupplier
     ) {
-      if (pathSupplier().length === 0) {
+      const range = new vscode.Range(
+        new vscode.Position(startLine, startCharacter),
+        new vscode.Position(startLine, startCharacter + length)
+      );
+      const jsonPathLen = pathSupplier().length;
+      if (jsonPathLen === 0) {
         if (property) {
           namespace = property + '::';
         }
-      } else if (pathSupplier().length === 2) {
+        tokensBuilder.push(range, 'namespace', ['declaration']);
+      } else if (jsonPathLen === 2) {
         etypeStart = new vscode.Position(startLine, startCharacter);
         etypeRange = new vscode.Range(
           new vscode.Position(startLine, startCharacter + 1),
           new vscode.Position(startLine, startCharacter + length - 1)
         );
+        deftype = property;
+        tokensBuilder.push(range, 'type', ['declaration']);
 
-        if (
-          pathSupplier()[1] === 'entityTypes' ||
-          pathSupplier()[1] === 'commonTypes'
-        ) {
+        if (pathSupplier()[1] === 'commonTypes') {
           etype = namespace + property;
+          symbol = vscode.SymbolKind.Struct;
+        } else if (pathSupplier()[1] === 'entityTypes') {
+          etype = namespace + property;
+          symbol = vscode.SymbolKind.Class;
         } else if (pathSupplier()[1] === 'actions') {
           etype = `${namespace}Action::"${property}"`;
+          symbol = vscode.SymbolKind.Function;
+        }
+      } else if (pathSupplier()[jsonPathLen - 1] === 'attributes') {
+        // anything directly under "attributes" is an property declaration
+        tokensBuilder.push(range, 'property', ['declaration']);
+      }
+    },
+    onLiteralValue(
+      value,
+      offset,
+      length,
+      startLine,
+      startCharacter,
+      pathSupplier
+    ) {
+      const jsonPathLen = pathSupplier().length;
+      const range = new vscode.Range(
+        new vscode.Position(startLine, startCharacter),
+        new vscode.Position(startLine, startCharacter + length)
+      );
+      if (
+        // anything directly under "memberOfTypes" under "entityTypes" is a type
+        jsonPathLen === 5 &&
+        pathSupplier()[1] === 'entityTypes' &&
+        pathSupplier()[3] === 'memberOfTypes'
+      ) {
+        tokensBuilder.push(range, 'type', []);
+        if (pathSupplier()[3] === 'memberOfTypes') {
+          definitions.push(
+            determineDefinitionRange(
+              value,
+              startLine,
+              startCharacter,
+              length - 1
+            )
+          );
+        }
+      } else if (
+        // anything directly under "principalTypes" or "resourceTypes" under "actions" is a type
+        jsonPathLen === 6 &&
+        pathSupplier()[1] === 'actions' &&
+        (pathSupplier()[4] === 'principalTypes' ||
+          pathSupplier()[4] === 'resourceTypes')
+      ) {
+        tokensBuilder.push(range, 'type', []);
+        definitions.push(
+          determineDefinitionRange(value, startLine, startCharacter, length - 1)
+        );
+      } else if (
+        // "id" or "type" under "memberOf" under "actions" is a type
+        jsonPathLen === 6 &&
+        pathSupplier()[1] === 'actions' &&
+        pathSupplier()[3] === 'memberOf' &&
+        (pathSupplier()[5] === 'id' || pathSupplier()[5] === 'type')
+      ) {
+        tokensBuilder.push(range, 'type', []);
+      } else if (pathSupplier()[jsonPathLen - 1] === 'name') {
+        // anything directly under "name" is (probably) a type
+        if (value === 'ipaddr' || value === 'decimal') {
+          tokensBuilder.push(range, 'function', []);
+        } else {
+          tokensBuilder.push(range, 'type', []);
+          definitions.push(
+            determineDefinitionRange(
+              value,
+              startLine,
+              startCharacter,
+              length - 1
+            )
+          );
         }
       }
     },
   });
 
-  schemaCache[schemaDoc.uri.toString()] = {
+  const cachedItem = {
     version: schemaDoc.version,
     entities: entities,
+    tokens: tokensBuilder.build(),
+    definitions: definitions,
   };
+  schemaCache[schemaDoc.uri.toString()] = cachedItem;
 
-  return entities;
+  return cachedItem;
+};
+
+export const schemaTokensProvider: vscode.DocumentSemanticTokensProvider = {
+  provideDocumentSemanticTokens(
+    cedarSchemaDoc: vscode.TextDocument
+  ): vscode.ProviderResult<vscode.SemanticTokens> {
+    return parseCedarDocSchema(cedarSchemaDoc).tokens;
+  },
 };