+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## v0.5.3 (Preview) 2023-10-09
+## v0.5.3 (Preview) 2023-10-16
 
 - Update to Cedar 2.4.0
-- Used Cedar policy `@id` value in Outline and Breadcrumb
+- Use Cedar policy `@id` value in Outline and Breadcrumb
+- Support "Go to Definition" on Cedar entity types 
 
 ## v0.5.2 (Preview) 2023-08-31